+++ packages/auth-zkp-iden3/src/index.ts
@@ -1,0 +1,2 @@
+export * from './auth-zkp'
+export * from './types'