+++ packages/shared-zkp-iden3/src/errors/runtime-errors.ts
@@ -1,0 +1,17 @@
+import { RuntimeError } from '@distributedlab/tools'
+
+export class IssuerResponseEmptyError extends RuntimeError {
+  public name = 'Issuer response empty'
+
+  constructor(originalError: Error | string) {
+    super(originalError)
+  }
+}
+
+export class FileEmptyError extends RuntimeError {
+  public name = 'Given file is empty'
+
+  constructor(originalError: Error | string) {
+    super(originalError)
+  }
+}