+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0-rc.3] - 2023-07-04
+### Added
+- `root` - `@rarimo/identity-gen-iden3` package
+- `root` - `@rarimo/auth-zkp-iden3` package
+- `root` - `@rarimo/zkp-gen-iden3` package
+- `root` - `@rarimo/shared-zkp-iden3` package
+- `@rarimo/shared` - export createApi method
 
 ## [2.0.0-rc.2] - 2023-06-13
 ### Fixed
@@ -305,7 +311,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.2...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.3...HEAD
+[2.0.0-rc.3]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.2...2.0.0-rc.3
 [2.0.0-rc.2]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.1...2.0.0-rc.2
 [2.0.0-rc.1]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.0...2.0.0-rc.1
 [2.0.0-rc.0]: https://github.com/rarimo/js-sdk/compare/1.5.0...2.0.0-rc.0